[~] mod create a section var interpreter for more organization and two differentes files
