+++ src/var_interpreters/var_interpreter.c
@@ -7,47 +7,6 @@
 
 #include "../../include/struct.h"
 
-char *simple(char c)
-{
-    char *str = malloc(sizeof(char) * 2);
-
-    if (!str)
-        return NULL;
-    str[0] = c;
-    str[1] = '\0';
-    return str;
-}
-
-char **array_null(char c)
-{
-    char **array = malloc(sizeof(char *) * (2));
-    char *str = NULL;
-
-    if (!array)
-        return NULL;
-    str = simple(c);
-    if (str == NULL) {
-        free(array);
-        return NULL;
-    }
-    array[0] = str;
-    array[1] = NULL;
-    return array;
-}
-
-static char *get_name(char *str, int max)
-{
-    char *res = malloc(sizeof(char) * (max + 2));
-
-    if (!res)
-        return NULL;
-    for (int i = 0; i != max + 1; i++) {
-        res[i] = str[i];
-    }
-    res[max + 1] = '\0';
-    return res;
-}
-
 static char *get_variable(char *str)
 {
     bool parenthesis = false;
@@ -70,24 +29,6 @@ static char *get_variable(char *str)
     return res;
 }
 
-char *my_strncat(char *base, char *value)
-{
-    int len = my_strlen(base);
-    int i = len;
-
-    if (!value)
-        return base;
-    base = realloc(base, sizeof(char) * (len + my_strlen(value) + 1));
-    if (!base)
-        return NULL;
-    for (int d = 0; value[d]; d++) {
-        base[i] = value[d];
-        i++;
-    }
-    base[i] = '\0';
-    return base;
-}
-
 static int replace_in_str(nodes_t *str, char *value, char *name, char **tmp)
 {
     char *res = NULL;
@@ -99,10 +40,10 @@ static int replace_in_str(nodes_t *str, char *value, char *name, char **tmp)
     }
     free(name);
     for (; tmp[i + 1] != NULL; i++) {
-        res = my_strncat(res, tmp[i]);
-        res = my_strncat(res, value);
+        res = my_str_rea(res, tmp[i]);
+        res = my_str_rea(res, value);
     }
-    res = my_strncat(res, tmp[i]);
+    res = my_str_rea(res, tmp[i]);
     free_array(tmp);
     if (!res)
         return ALTERNATIVE_EXIT;
@@ -148,23 +89,6 @@ static int search_var(tcsh_t *term, nodes_t *str, char *var, char *name)
     return ALTERNATIVE_EXIT;
 }
 
-char *without(char *str)
-{
-    char *res = malloc(sizeof(char) * (my_strlen(str)));
-    int j = 0;
-
-    if (!res)
-        return str;
-    for (int i = 0; str[i]; i++) {
-        if (str[i] == '{' || str[i] == '}')
-            continue;
-        res[j] = str[i];
-        j++;
-    }
-    res[j] = '\0';
-    return res;
-}
-
 static int interpreter(tcsh_t *term, nodes_t *str, int i)
 {
     char *name = NULL;
@@ -178,18 +102,21 @@ static int interpreter(tcsh_t *term, nodes_t *str, int i)
     return SUCCESS_EXIT;
 }
 
+static int check(tcsh_t *term, nodes_t *str, int i)
+{
+    return (interpreter(term, str, i) != SUCCESS_EXIT ? -1 : 0);
+}
+
 int search_variable(tcsh_t *term, nodes_t *str)
 {
     int i = 0;
 
     while (((char *)str->data)[i]) {
-        if (((char *)str->data)[i] == '$') {
-            if (interpreter(term, str, i) != SUCCESS_EXIT)
-                return ALTERNATIVE_EXIT;
-            i = 0;
+        i = ((char *)str->data)[i] == '$' ? check(term, str, i) : i + 1;
+        if (i == 0)
             continue;
-        }
-        i++;
+        if (i == -1)
+            return ALTERNATIVE_EXIT;
     }
     return SUCCESS_EXIT;
 }
